% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\docType{data}
\name{accident_2013}
\alias{accident_2013}
\title{Accidents in US National Highway Traffic 2013}
\format{A .csv compressed in .bz2 format to read with \code{\link{read_csv}} from package \code{readr}. 
\describe{
50 variables and 30202 observations
}}
\source{
This is data from the US National Highway Traffic Safety Administration's Fatality Analysis Reporting System, which is a nationwide census providing the American public yearly data regarding fatal injuries suffered in motor vehicle traffic crashes in 2013.
}
\usage{
accident_2013
}
\description{
Accidents in US National Highway Traffic 2013
}
\examples{
\dontrun{
 accident_2013
}
}
\keyword{datasets}
