% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\docType{data}
\name{accident_2015}
\alias{accident_2015}
\title{Accidents in US National Highway Traffic 2015}
\format{A .csv compressed in .bz2 format to read with \code{\link{read_csv}} from package \code{readr}. 
\describe{
50 variables and 32166 observations
}}
\source{
This is data from the US National Highway Traffic Safety Administration's Fatality Analysis Reporting System, which is a nationwide census providing the American public yearly data regarding fatal injuries suffered in motor vehicle traffic crashes in 2015.
}
\usage{
accident_2015
}
\description{
Accidents in US National Highway Traffic 2015
}
\examples{
\dontrun{
 accident_2015
}
}
\keyword{datasets}
