% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSDR_v1.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{eq_clean_data}
\usage{
eq_clean_data(data)
}
\arguments{
\item{data}{A data frame with raw data from NOAA (see details)}
}
\value{
data A data frame with cleaned columns
}
\description{
Clean up the format of certain columns in a NOAA dataframe
}
\details{
This function takes raw NOAA data frame (you can download a sample from \href{https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1}{here})
and returns a clean data frame with:
- A DATE column created by uniting the year, month, day and converting it to the Date class
- LATITUDE and LONGITUDE columns converted to numeric class
- LOCATION_NAME is modified removing the country name from it and changed to title case (see \code{\link{eq_location_clean}})
}
\examples{
\dontrun{
data <- readr::read_delim("earthquakes.tsv.gz", delim = "\\t")
clean_data <- eq_clean_data(data)
}

}
