% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{eq_clean_date}
\alias{eq_clean_date}
\title{Clean earthquak dates}
\usage{
eq_clean_date(df)
}
\arguments{
\item{df}{A data.frame such as this obtainded from NOAA.}
}
\value{
A data.frame with a tidy date column.
}
\description{
Clean earthquak dates
}
