% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{eq_clean_location}
\alias{eq_clean_location}
\title{Clean earthquak locations}
\usage{
eq_clean_location(df)
}
\arguments{
\item{df}{A data.frame such as this obtainded from NOAA.}
}
\value{
A data.fram with a tidy location column
}
\description{
Clean earthquak locations
}
