% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Create earthquak label}
\usage{
eq_create_label(df)
}
\arguments{
\item{df}{A data.frame such as this obtainded from NOAA.}
}
\value{
A column of earthquak annotation; locaion, magnitude and deaths. Ignores the fiels when NA.
}
\description{
Create earthquak label
}
