% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSDR_v1.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{eq_location_clean}
\usage{
eq_location_clean(data)
}
\arguments{
\item{data}{A data frame with raw data from NOAA (see details)}
}
\value{
data A data frame with cleaned LOCATION_NAME format
}
\description{
Clean up the format of LOCATION_NAME column in a NOAA dataframe (see \code{\link{eq_clean_data}} for more details)
}
\details{
cleans the LOCATION_NAME column by stripping out the country name (including the colon) and converts names
to title case (as opposed to all caps). This function is only used inside \code{\link{eq_clean_data}}, not exported.
}
\examples{
\dontrun{
data <- readr::read_delim("earthquakes.tsv.gz", delim = "\\t")
clean_data <- eq_location_clean(data)
}


}
