% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{eq_map}
\alias{eq_map}
\title{Map earthquak data to a leaflet tile}
\usage{
eq_map(df)
}
\arguments{
\item{df}{A data.frame such as this obtainded from NOAA.}
}
\value{
A graph with markers for specified earthquaks and popups with info; location, magnitude and deaths.
}
\description{
Map earthquak data to a leaflet tile
}
