% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Reads .csv data and plots a map of states}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{a single value of US state number. eg: 1, 10 or 12;}

\item{year}{a single value representing a year. eg. 2013, 2014}
}
\value{
a plot of a US State.
}
\description{
This function to read a .csv extension file and returns a plot of the State map. The user must pass the state number and a year of data fot that.
}
\details{
This function along with \code{\link{make_filename}} and \code{\link{fars_read}} reads data from .csv types only receiving as param a value of state number and year and plot a map of the input State. This function invokes the functions \code{\link{filter}} from package \code{dplyr} and \code{\link{map}} from \code{maps} also as \code{\link{points}} from \code{graphics}.
}
\note{
if any of those packages where not instaled or the arguments passed are wrong no plot ir shown.
}
\examples{
\dontrun{
state.num <- 10; year <- 2013
fars_map_state(state.num, year)
}

}
