% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Reads data from a file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{text string containing the full path and name of the file to be read.}
}
\value{
a tible, table, data.frame data object
}
\description{
This is a simple function to read data from .csv extension and return a tible in R.
}
\details{
this function read data using the function \code{\link{read_csv}} from package \code{readr}. This function is capable to read \code{csv} files compressed in .bz2 format. The next step is convert the data into a tible format by using the function \code{\link{tbl_df}} from package \code{dplyr}. If the full path is not well passed the function may crash.
}
\examples{

path <- "accident_2013.csv.bz2"
x <- fars_read(path)
class(x)
head(x)

}
