% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Reads data from several files}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{vector or list of values repsenting years: eg c(2013, 2014) or list(2013, 2014).}
}
\value{
A lis of a tible objects
}
\description{
This function to read several .csv extension files and returns a list of tidy files containing only Month and Year inside of each.
}
\details{
This function along with \code{\link{make_filename}}, \code{\link{fars_read}} reads data from .csv types only receiving as param a vector or list of years by the user and returns a list of tibles containing variables YEAR and MONTH present inside those data. The function \code{\link{mutate}} and \code{\link{select}} from package \code{dplyr} are used to tidy data.
}
\examples{

year <- c(2013, 2014)
x <- fars_read_years(year)
lapply(x, function(x) head(x, 2))

}
