% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Reads data from several files and returns the row number of each one}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{vector or list of values repsenting years: eg c(2013, 2014) or list(2013, 2014).}
}
\value{
A tible containg the rown umbers of each dataset by month and year.
}
\description{
This function to read several .csv extension files and returns a tible where each column represents a year and each row represents a month. The values column vs row are the row number of each dataset/year.
}
\details{
This function along with \code{\link{fars_read_years}} reads data from .csv types only receiving as param a vector or list of years by the user and returns a tible containing the row number by month and year where each row is a month and eacho column is a year passed by the user and present in the data sets. This function invokes the functions \code{\link{bind_rows}}, \code{\link{group_by}} and \code{\link{summarize}} from package \code{dplyr} and \code{\link{spread}} from \code{tidyr}.
}
\examples{
years <- c(2013, 2014, 2015)
x <- fars_summarize_years(years)
class(x)
head(x)

}
