% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{Timeline label}
\usage{
geom_timeline_label(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = FALSE,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{...}{}
}
\value{
Labels to use with geom_timeline
}
\description{
Timeline label
}
