% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSDR_v1.R
\docType{data}
\name{geomtimeline}
\alias{geomtimeline}
\title{geomtimeline}
\format{An object of class \code{Geomtimeline} (inherits from \code{Geom}, \code{ggproto}) of length 6.}
\usage{
geomtimeline
}
\arguments{
\item{xmaxdate}{date or charachter formatted in year, month, day order}

\item{xmindate}{date or charachter formatted in year, month, day order}
}
\description{
Creates the grob for \code{\link{geom_timeline}}
}
\keyword{datasets}
