% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSDR_v1.R
\docType{data}
\name{geomtimeline_label}
\alias{geomtimeline_label}
\title{geomtimeline_label}
\format{An object of class \code{Geomtimeline_label} (inherits from \code{Geom}, \code{ggproto}) of length 6.}
\usage{
geomtimeline_label
}
\description{
Creates the grob for \code{\link{geom_timeline_label}}
}
\keyword{datasets}
