% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create correct path to read data from a file}
\usage{
make_filename(year)
}
\arguments{
\item{year}{text string containing value of the year in format "2013" being text or number.}
}
\value{
a string describing the full correct path to the file to be read.
}
\description{
This is a simple function to prepare a full correct path to the function \code{\link{fars_read}}.
}
\details{
the value from this function intends to be used to the function \code{\link{fars_read}} since it require a well formated full path from the file to be read. If the file to be read is not in the correct directory the string will not be good to the function \code{\link{fars_read}}.
}
\note{
make sure that the file to read is in the main directory before run this function along with \code{\link{fars_read}}.
}
\examples{

path <- "2013"
x <- make_filename(path)
class(x)
head(x)
head(fars_read(x))

}
