% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{top_eq}
\alias{top_eq}
\title{Filter earthquakes by magnitude}
\usage{
top_eq(data, n_max = Inf, min_date = -Inf, max_date = Inf)
}
\arguments{
\item{data}{Earthquake data}

\item{n_max}{Max number of earthquakes to pick}

\item{min_date}{Minimum date}

\item{max_date}{Maximum date}
}
\value{
A dataframe with the filtered data
}
\description{
Filter the top n_max earthquakes according to their magnitude.
}
\examples{
top_10 <- top_earthquakes(earthquake_data, 10)

}
