% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSDR_v1.R
\name{eq_create_label}
\alias{eq_create_label}
\title{eq_create_label}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{a dataframe containing NOAA details}
}
\value{
character string
}
\description{
Creates a html string for a more representative label.
}
\details{
put together a character string for each earthquake that shows the cleaned location (as cleaned by the \code{\link{eq_location_clean}}),
the magnitude (EQ_PRIMARY), and the total number of deaths (TOTAL_DEATHS), with boldface labels for each ("Location","Total deaths",
and "Magnitude"). If an earthquake is missing values for any of these, both the label and the value are skipped for that element of the tag.
}
\examples{
\dontrun{
readr::read_delim("earthquakes.tsv.gz", delim = "\\t") \%>\%
eq_clean_data() \%>\%
dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
dplyr::mutate(popup_text = eq_create_label(.)) \%>\%
eq_map(annot_col = "popup_text")
}

}
