% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSDR_v1.R
\name{eq_map}
\alias{eq_map}
\title{eq_map}
\usage{
eq_map(data, annot_col = NULL)
}
\arguments{
\item{data}{a dataframe containing NOAA details}

\item{annot_col}{any column of the dataframe, will be converted to character and used as label}
}
\value{
a leaflet map
}
\description{
Creates a html map using leaflet to show NOAA data (see \code{\link{eq_clean_data}} for more details)
}
\details{
The function maps the epicenters (LATITUDE/LONGITUDE) and annotates each point with in pop up window containing annotation data
stored in a column of the data frame. The user should be able to choose which column is used for the annotation in the pop-up with
a function argument named annot_col. Each earthquake should be shown with a circle, and the radius of the circle should be proportional
to the earthquake's magnitude (EQ_PRIMARY).
}
\examples{
\dontrun{
readr::read_delim("earthquakes.tsv.gz", delim = "\\t") \%>\%
eq_clean_data() \%>\%
dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
eq_map(annot_col = "DATE")
}

}
