% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSDR_v1.R
\name{geom_timeline}
\alias{geom_timeline}
\title{geom_timeline}
\usage{
geom_timeline(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link{aes}} or
\code{\link{aes_}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{a dataframe containing NOAA details}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link{borders}}.}

\item{...}{other arguments passed on to \code{\link{layer}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{color = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}
}
\description{
Creates a timeline chart specific for a NOAA dataframe (see \code{\link{eq_clean_data}} for more details)
}
\details{
plotting a time line of earthquakes ranging from xmin to xmaxdates with a point for each earthquake.
Optional aesthetics include color, size, and alpha (for transparency). The x aesthetic is a date and an optional y aesthetic
is a factor indicating some stratification in which case multiple time lines will be plotted for each level of the factor (e.g. country).
}
\examples{
\dontrun{
data \%>\%
 filter(COUNTRY \%in\% c("USA", "CHINA")) \%>\%
 eq_clean_data()\%>\%
 ggplot(aes(x = DATE, y = COUNTRY, alpha = TOTAL_DEATHS, size = EQ_PRIMARY))+
 geom_timeline(xmindate = "2008/01/01", xmaxdate = "2015/12/31")
}

}
