% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSDR_v1.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{geom_timeline_label}
\usage{
geom_timeline_label(mapping = NULL, stat = "identity", data = NULL,
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link{aes}} or
\code{\link{aes_}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{data}{a dataframe containing NOAA details}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link{borders}}.}

\item{...}{other arguments passed on to \code{\link{layer}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{color = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{n_max}{numeric, it requires a value defined for the aestetic size, if omitted is by default = 5}
}
\description{
Creates additional notation for the timeline chart (see \code{\link{geom_timeline}} for more details)
}
\details{
This geom adds a vertical line to each data point with a text annotation (e.g. the location of the earthquake) attached to each line.
There is an option to subset to n_max number of earthquakes, where we take the n_max largest (by magnitude) earthquakes.
}
\examples{
\dontrun{
data \%>\%
 filter(COUNTRY \%in\% c("USA", "CHINA")) \%>\%
 eq_clean_data()\%>\%
 ggplot(aes(x = DATE, y = COUNTRY, alpha = TOTAL_DEATHS, size = EQ_PRIMARY))+
 geom_timeline()+
 geom_timeline_lable(aes(label = LOCATION_NAME),n_max = 3)
}

}
